requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
pandas==1.3.5
SQLAlchemy==1.4.34
psycopg2-binary==2.9.3
openpyxl==latest
lxml==latest